Use BigDecimal and BigInteger instead of int, Integer, double, Double

+++ src/main/java/wild/yellow/FileReader.java
@@ -4,14 +4,16 @@ import lombok.Getter;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
 @Getter
 public class FileReader {
 
-    private final List<Integer> integers = new ArrayList<>();
-    private final List<Double> doubles = new ArrayList<>();
+    private final List<BigInteger> integers = new ArrayList<>();
+    private final List<BigDecimal> doubles = new ArrayList<>();
     private final List<String> strings = new ArrayList<>();
 
     Option option;
@@ -29,9 +31,9 @@ public class FileReader {
                         DataTypeDefiner.DefinedType type = DataTypeDefiner.define(line.trim());
                         switch (type.type()) {
                             case INTEGER ->
-                                    integers.add((Integer) type.value());
+                                    integers.add((BigInteger) type.value());
                             case DOUBLE ->
-                                    doubles.add((Double) type.value());
+                                    doubles.add((BigDecimal) type.value());
                             case STRING ->
                                     strings.add((String) type.value());
                         }
